feat(avo-2599): Add share dropdown to assignment edit page
Refactor share assignment with pupils component to a reusable shared component, add share with pupils component to sharedropdown
