+++ src/shared/components/ShareWithPupils/ShareWithPupils.scss
@@ -1,4 +1,4 @@
-@import '../../styles/settings/variables';
+@import '../../../styles/settings/variables';
 
 .c-share-assignment-with-pupil-popup {
 	padding: 1.5rem;
